PUB-1686 Refactor publication service (#258)
* WIP

* refactor publicationservice class

* refactor code and fix PMD issues

* fix build

* refactor code

* fix build

* fix code review comment

* PUB-1634 Added functional tests for "Unauthorized" status code (#257)

* PUB-1678 Java 17 upgrade (#236)

* Upgraded to Java 17

* Updated app insights in docker and further gradle updates

* Updated PMD and Checkstyles

* PUB-1678 - Fixed sonarqube

* PUB-1678 - Updated ruleset

* Updated suppressions

* Vulns


* Updated Suppressions (#238)

* Updated Suppressions

* Updated suppressions

* Updated IaC for apim app insights (#234)


* PUB-1625 Added custom logging for failed publications (#235)

* Added custom logs for failed publications

* Null checks for dates

* Added functional tests for "Unauthorized" status code


* fix code review comments and conflicts
