+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/service/PublicationService.java
@@ -2,32 +2,24 @@ package uk.gov.hmcts.reform.pip.data.management.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.Resource;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import uk.gov.hmcts.reform.pip.data.management.database.ArtefactRepository;
 import uk.gov.hmcts.reform.pip.data.management.database.AzureBlobService;
 import uk.gov.hmcts.reform.pip.data.management.database.LocationRepository;
-import uk.gov.hmcts.reform.pip.data.management.errorhandling.exceptions.ArtefactNotFoundException;
+import uk.gov.hmcts.reform.pip.data.management.helpers.ArtefactHelper;
 import uk.gov.hmcts.reform.pip.data.management.models.location.Location;
-import uk.gov.hmcts.reform.pip.data.management.models.location.LocationType;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Artefact;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
-import uk.gov.hmcts.reform.pip.data.management.models.publication.Sensitivity;
-import uk.gov.hmcts.reform.pip.data.management.utils.CaseSearchTerm;
+import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactTriggerService;
 import uk.gov.hmcts.reform.pip.data.management.utils.PayloadExtractor;
 import uk.gov.hmcts.reform.pip.model.enums.UserActions;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
 
 import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
 
@@ -37,7 +29,6 @@ import static uk.gov.hmcts.reform.pip.model.LogBuilder.writeLog;
 
 @Slf4j
 @Service
-@SuppressWarnings({"PMD.GodClass", "PMD.LawOfDemeter"})
 public class PublicationService {
 
     private static final char DELIMITER = ',';
@@ -48,33 +39,25 @@ public class PublicationService {
 
     private final PayloadExtractor payloadExtractor;
 
-    private final SubscriptionManagementService subscriptionManagementService;
-
     private final LocationRepository locationRepository;
 
-    private final AccountManagementService accountManagementService;
-
-    private final PublicationServicesService publicationServicesService;
-
     private final ChannelManagementService channelManagementService;
+
+    private final ArtefactTriggerService artefactTriggerService;
 
     @Autowired
     public PublicationService(ArtefactRepository artefactRepository,
                               AzureBlobService azureBlobService,
                               PayloadExtractor payloadExtractor,
-                              SubscriptionManagementService subscriptionManagementService,
-                              AccountManagementService accountManagementService,
                               LocationRepository locationRepository,
-                              PublicationServicesService publicationServicesService,
-                              ChannelManagementService channelManagementService) {
+                              ChannelManagementService channelManagementService,
+                              ArtefactTriggerService artefactTriggerService) {
         this.artefactRepository = artefactRepository;
         this.azureBlobService = azureBlobService;
         this.payloadExtractor = payloadExtractor;
-        this.subscriptionManagementService = subscriptionManagementService;
-        this.accountManagementService = accountManagementService;
         this.locationRepository = locationRepository;
-        this.publicationServicesService = publicationServicesService;
         this.channelManagementService = channelManagementService;
+        this.artefactTriggerService = artefactTriggerService;
     }
 
     /**
@@ -95,13 +78,13 @@ public class PublicationService {
         boolean isExisting = applyExistingArtefact(artefact);
 
         String blobUrl = azureBlobService.createPayload(
-            isExisting ? getUuidFromUrl(artefact.getPayload()) : UUID.randomUUID().toString(),
+            isExisting ? ArtefactHelper.getUuidFromUrl(artefact.getPayload()) : UUID.randomUUID().toString(),
             payload
         );
 
         // Add 7 days to the expiry date if the list type is SJP
-        if (ListType.SJP_PUBLIC_LIST.equals(artefact.getListType())
-            || ListType.SJP_PRESS_LIST.equals(artefact.getListType())) {
+        if (artefact.getListType().equals(ListType.SJP_PUBLIC_LIST)
+            || artefact.getListType().equals(ListType.SJP_PRESS_LIST)) {
             artefact.setExpiryDate(artefact.getExpiryDate().plusDays(7));
         }
 
@@ -126,7 +109,7 @@ public class PublicationService {
         boolean isExisting = applyExistingArtefact(artefact);
 
         String blobUrl = azureBlobService.uploadFlatFile(
-            isExisting ? getUuidFromUrl(artefact.getPayload()) : UUID.randomUUID().toString(),
+            isExisting ? ArtefactHelper.getUuidFromUrl(artefact.getPayload()) : UUID.randomUUID().toString(),
             file
         );
 
@@ -142,7 +125,7 @@ public class PublicationService {
     @Async
     public void processCreatedPublication(Artefact artefact) {
         channelManagementService.requestFileGeneration(artefact.getArtefactId());
-        checkAndTriggerSubscriptionManagement(artefact);
+        artefactTriggerService.checkAndTriggerSubscriptionManagement(artefact);
     }
 
     /**
@@ -168,266 +151,6 @@ public class PublicationService {
         return foundArtefact.isPresent();
     }
 
-    private String getUuidFromUrl(String payloadUrl) {
-        return payloadUrl.substring(payloadUrl.lastIndexOf('/') + 1);
-    }
-
-
-    /**
-     * Get all relevant artefacts relating to a given location ID.
-     *
-     * @param searchValue - represents the location ID in question being searched for
-     * @param userId      - represents the user ID of the user who is making the request
-     * @return a list of all artefacts that fulfil the timing criteria, match the given location id and sensitivity
-     *     associated with given verification status.
-     */
-    public List<Artefact> findAllByLocationId(String searchValue, UUID userId) {
-        LocalDateTime currDate = LocalDateTime.now();
-        List<Artefact> artefacts = artefactRepository.findArtefactsByLocationId(searchValue, currDate);
-
-        return artefacts.stream().filter(artefact -> isAuthorised(artefact, userId)).toList();
-    }
-
-    /**
-     * Get all artefacts for admin actions.
-     *
-     * @param locationId The location id to search for.
-     * @param userId     represents the user ID of the user who is making the request
-     * @param isAdmin    bool to check whether admin search is needed, if not will default to findAllByLocationId().
-     * @return list of matching artefacts.
-     */
-    public List<Artefact> findAllByLocationIdAdmin(String locationId, UUID userId, boolean isAdmin) {
-        log.info(writeLog("ADMIN - Searching for all artefacts with " + locationId));
-        return isAdmin
-            ? artefactRepository.findArtefactsByLocationIdAdmin(locationId) : findAllByLocationId(locationId, userId);
-    }
-
-    /**
-     * Get all relevant Artefacts based on search values stored in the Artefact.
-     *
-     * @param searchTerm  the search term checking against, eg. CASE_ID or CASE_URN
-     * @param searchValue the search value to look for
-     * @param userId      represents the user ID of the user who is making the request
-     * @return list of Artefacts
-     */
-    public List<Artefact> findAllBySearch(CaseSearchTerm searchTerm, String searchValue, UUID userId) {
-        LocalDateTime currDate = LocalDateTime.now();
-        List<Artefact> artefacts;
-        switch (searchTerm) {
-            case CASE_ID, CASE_URN ->
-                artefacts = artefactRepository.findArtefactBySearch(searchTerm.dbValue, searchValue, currDate);
-            case CASE_NAME -> artefacts = artefactRepository.findArtefactByCaseName(searchValue, currDate);
-            default -> throw new IllegalArgumentException(String.format("Invalid search term: %s", searchTerm));
-        }
-
-        artefacts = artefacts.stream().filter(artefact -> isAuthorised(artefact, userId)).toList();
-
-        if (artefacts.isEmpty()) {
-            throw new ArtefactNotFoundException(String.format("No Artefacts found with for %s with the value: %s",
-                                                              searchTerm, searchValue
-            ));
-        }
-        return artefacts;
-    }
-
-    public Artefact getMetadataByArtefactId(UUID artefactId) {
-        return artefactRepository.findArtefactByArtefactId(artefactId.toString())
-            .orElseThrow(() -> new ArtefactNotFoundException(String.format(
-                "No artefact found with the ID: %s",
-                artefactId
-            )));
-    }
-
-    /**
-     * Takes in artefact id and returns the metadata for the artefact.
-     *
-     * @param artefactId represents the artefact id which is then used to get an artefact to populate the inputs
-     *                   for the blob request.
-     * @param userId     represents the user ID of the user who is making the request
-     * @return The metadata for the found artefact.
-     */
-    public Artefact getMetadataByArtefactId(UUID artefactId, UUID userId) {
-
-        LocalDateTime currentDate = LocalDateTime.now();
-
-        Optional<Artefact> artefact = artefactRepository.findByArtefactId(
-            artefactId.toString(),
-            currentDate
-        );
-
-        if (artefact.isPresent() && isAuthorised(artefact.get(), userId)) {
-            return artefact.get();
-        }
-
-        throw new ArtefactNotFoundException(String.format("No artefact found with the ID: %s", artefactId));
-    }
-
-    /**
-     * Takes in artefact id and returns the payload within the matching blob in string format.
-     *
-     * @param artefactId represents the artefact id which is then used to get an artefact to populate the inputs
-     *                   for the blob request.
-     * @param userId     represents the user ID of the user who is making the request
-     * @return The data within the blob in string format.
-     */
-    public String getPayloadByArtefactId(UUID artefactId, UUID userId) {
-        Artefact artefact = getMetadataByArtefactId(artefactId, userId);
-
-        return azureBlobService.getBlobData(getUuidFromUrl(artefact.getPayload()));
-    }
-
-    /**
-     * Takes in artefact id and returns the payload within the matching blob in string format. This is used for admin
-     * requests
-     *
-     * @param artefactId represents the artefact id which is then used to get an artefact to populate the inputs
-     *                   for the blob request.
-     * @return The data within the blob in string format.
-     */
-    public String getPayloadByArtefactId(UUID artefactId) {
-        Artefact artefact = getMetadataByArtefactId(artefactId);
-
-        return azureBlobService.getBlobData(getUuidFromUrl(artefact.getPayload()));
-    }
-
-    /**
-     * Retrieves a flat file for an artefact.
-     *
-     * @param artefactId The artefact ID to retrieve the flat file from.
-     * @param userId     represents the user ID of the user who is making the request
-     * @return The flat file resource.
-     */
-    public Resource getFlatFileByArtefactID(UUID artefactId, UUID userId) {
-        Artefact artefact = getMetadataByArtefactId(artefactId, userId);
-
-        return azureBlobService.getBlobFile(getUuidFromUrl(artefact.getPayload()));
-    }
-
-    /**
-     * Retrieves a flat file for an artefact. This is used for admin requests
-     *
-     * @param artefactId The artefact ID to retrieve the flat file from.
-     * @return The flat file resource.
-     */
-    public Resource getFlatFileByArtefactID(UUID artefactId) {
-        Artefact artefact = getMetadataByArtefactId(artefactId);
-
-        return azureBlobService.getBlobFile(getUuidFromUrl(artefact.getPayload()));
-    }
-
-    /**
-     * Checks if the artefact has a display from date of today or previous then triggers the sub fulfilment
-     * process on subscription-management if appropriate.
-     */
-    public void checkAndTriggerSubscriptionManagement(Artefact artefact) {
-        //TODO: fully switch this logic to localdates once artefact model changes //NOSONAR
-        if (artefact.getDisplayFrom().toLocalDate().isBefore(LocalDate.now().plusDays(1))
-            && (artefact.getDisplayTo() == null
-            || artefact.getDisplayTo().toLocalDate().isAfter(LocalDate.now().minusDays(1)))) {
-            log.info(sendArtefactForSubscription(artefact));
-        }
-    }
-
-    public String sendArtefactForSubscription(Artefact artefact) {
-        return subscriptionManagementService.sendArtefactForSubscription(artefact);
-    }
-
-    /**
-     * Scheduled method that checks daily for newly dated from artefacts.
-     */
-    public void checkNewlyActiveArtefacts() {
-        artefactRepository.findArtefactsByDisplayFrom(LocalDate.now())
-            .forEach(artefact -> log.info(sendArtefactForSubscription(artefact)));
-    }
-
-    /**
-     * Find artefacts with NoMatch location and send them for reporting.
-     */
-    public void reportNoMatchArtefacts() {
-        findNoMatchArtefactsForReporting(artefactRepository.findAllNoMatchArtefacts());
-    }
-
-    /**
-     * Attempts to delete an artefact along with the stored blobs.
-     *
-     * @param artefactId The ID of the artefact to be deleted.
-     * @param issuerId   The ID of the admin user who is attempting to delete the artefact.
-     */
-    public void deleteArtefactById(String artefactId, String issuerId) {
-        Optional<Artefact> artefactToDelete = artefactRepository.findArtefactByArtefactId(artefactId);
-        if (artefactToDelete.isPresent()) {
-            deleteAllPublicationBlobData(artefactToDelete.get());
-            artefactRepository.delete(artefactToDelete.get());
-            log.info(writeLog(issuerId, UserActions.REMOVE, artefactId));
-            triggerThirdPartyArtefactDeleted(artefactToDelete.get());
-        } else {
-            throw new ArtefactNotFoundException("No artefact found with the ID: " + artefactId);
-        }
-    }
-
-    /**
-     * Method that handles the logic to archive an artefact and delete the stored blobs.
-     *
-     * @param artefactId The ID of the artefact to be deleted.
-     * @param issuerId   The ID of the admin user who is attempting to delete the artefact.
-     */
-    @Transactional
-    public void archiveArtefactById(String artefactId, String issuerId) {
-        Optional<Artefact> artefactToArchive = artefactRepository.findArtefactByArtefactId(artefactId);
-
-        if (artefactToArchive.isPresent()) {
-            deleteAllPublicationBlobData(artefactToArchive.get());
-            artefactRepository.archiveArtefact(artefactId);
-            log.info(String.format("Artefact archived by %s, with artefact id: %s", issuerId, artefactId));
-            triggerThirdPartyArtefactDeleted(artefactToArchive.get());
-        } else {
-            throw new ArtefactNotFoundException("No artefact found with the ID: " + artefactId);
-        }
-    }
-
-    /**
-     * Archive expired artefacts from the database, Artefact and Publications Azure storage.
-     */
-    @Transactional
-    public void archiveExpiredArtefacts() {
-        LocalDateTime searchDateTime = LocalDateTime.now();
-        List<Artefact> outdatedArtefacts = artefactRepository.findOutdatedArtefacts(searchDateTime);
-
-        outdatedArtefacts.forEach(artefact -> {
-            artefactRepository.archiveArtefact(artefact.getArtefactId().toString());
-            deleteAllPublicationBlobData(artefact);
-        });
-
-        log.info(writeLog(String.format("%s outdated artefacts found and archived for before %s",
-                                        outdatedArtefacts.size(), searchDateTime
-        )));
-    }
-
-    /**
-     * Delete all data stored in the blobstore for an artefact.
-     *
-     * @param artefact The artefact requiring blob deletion.
-     */
-    private void deleteAllPublicationBlobData(Artefact artefact) {
-        // Delete the payload/flat file from the publications store
-        azureBlobService.deleteBlob(getUuidFromUrl(artefact.getPayload()));
-
-        // Try to delete the generated files for the publications if it's not a flat file
-        if (Boolean.FALSE.equals(artefact.getIsFlatFile())) {
-            try {
-                azureBlobService.deletePublicationBlob(artefact.getArtefactId() + ".pdf");
-
-                // If it's an SJP list the xlsx file also needs to be deleted
-                if (ListType.SJP_PUBLIC_LIST.equals(artefact.getListType())
-                    || ListType.SJP_PRESS_LIST.equals(artefact.getListType())) {
-                    azureBlobService.deletePublicationBlob(artefact.getArtefactId() + ".xlsx");
-                }
-            } catch (Exception ex) {
-                log.info("Failed to delete the generated publication file. Message: " + ex.getMessage());
-            }
-        }
-    }
-
     private void applyInternalLocationId(Artefact artefact) {
         if ("MANUAL_UPLOAD".equalsIgnoreCase(artefact.getProvenance())) {
             return;
@@ -444,60 +167,6 @@ public class PublicationService {
         } else {
             artefact.setLocationId(String.format("NoMatch%s", artefact.getLocationId()));
         }
-    }
-
-    private boolean isAuthorised(Artefact artefact, UUID userId) {
-        if (artefact.getSensitivity().equals(Sensitivity.PUBLIC)) {
-            return true;
-        } else if (userId == null) {
-            return false;
-        } else {
-            return accountManagementService.getIsAuthorised(userId, artefact.getListType(), artefact.getSensitivity());
-        }
-    }
-
-    public LocationType getLocationType(ListType listType) {
-        return listType.getListLocationLevel();
-    }
-
-    /**
-     * Triggers subscription management to handle deleted artefact to third party subscribers.
-     *
-     * @param deletedArtefact deleted artefact to notify of.
-     */
-    private void triggerThirdPartyArtefactDeleted(Artefact deletedArtefact) {
-        log.info(writeLog(subscriptionManagementService.sendDeletedArtefactForThirdParties(deletedArtefact)));
-    }
-
-    /**
-     * Receives a list of no match artefacts, checks it's not empty and create a map of location id to Provenance.
-     * Send this on to publication services.
-     *
-     * @param artefactList A list of no match artefacts
-     */
-    private void findNoMatchArtefactsForReporting(List<Artefact> artefactList) {
-        if (!artefactList.isEmpty()) {
-            Map<String, String> locationIdProvenanceMap = new ConcurrentHashMap<>();
-            artefactList.forEach(artefact -> locationIdProvenanceMap.put(
-                artefact.getLocationId().split("NoMatch")[1], artefact.getProvenance()));
-
-            log.info(publicationServicesService.sendNoMatchArtefactsForReporting(locationIdProvenanceMap));
-        }
-    }
-
-    /**
-     * Returns what is essentially a CSV file with the count of artefacts in a given location.
-     *
-     * @return string representing the csv file.
-     */
-    public String countArtefactsByLocation() {
-        List<String> returnedData = artefactRepository.countArtefactsByLocation();
-        StringBuilder builder = new StringBuilder(150);
-        builder.append("location,count\n");
-        for (String s : returnedData) {
-            builder.append(s).append('\n');
-        }
-        return builder.toString();
     }
 
     /**
