+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/controllers/PublicationControllerTest.java
@@ -22,6 +22,10 @@ import uk.gov.hmcts.reform.pip.data.management.models.publication.ListType;
 import uk.gov.hmcts.reform.pip.data.management.models.publication.Sensitivity;
 import uk.gov.hmcts.reform.pip.data.management.service.PublicationService;
 import uk.gov.hmcts.reform.pip.data.management.service.ValidationService;
+import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactDeleteService;
+import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactSearchService;
+import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactService;
+import uk.gov.hmcts.reform.pip.data.management.service.artefact.ArtefactTriggerService;
 import uk.gov.hmcts.reform.pip.data.management.utils.CaseSearchTerm;
 
 import java.io.IOException;
@@ -49,6 +53,18 @@ class PublicationControllerTest {
 
     @Mock
     private PublicationService publicationService;
+
+    @Mock
+    private ArtefactSearchService artefactSearchService;
+
+    @Mock
+    private ArtefactService artefactService;
+
+    @Mock
+    private ArtefactDeleteService artefactDeleteService;
+
+    @Mock
+    private ArtefactTriggerService artefactTriggerService;
 
     @Mock
     private ValidationService validationService;
@@ -171,7 +187,8 @@ class PublicationControllerTest {
 
     @Test
     void testGetArtefactsBySearchReturnsWhenTrue() {
-        when(publicationService.findAllBySearch(SEARCH_TERM, TEST_STRING, USER_ID)).thenReturn(List.of(artefactWithId));
+        when(artefactSearchService.findAllBySearch(SEARCH_TERM, TEST_STRING, USER_ID))
+            .thenReturn(List.of(artefactWithId));
         assertEquals(HttpStatus.OK, publicationController.getAllRelevantArtefactsBySearchValue(SEARCH_TERM, TEST_STRING,
                                                                                                USER_ID).getStatusCode(),
                      STATUS_CODE_MATCH);
@@ -179,15 +196,17 @@ class PublicationControllerTest {
 
     @Test
     void testGetArtefactsBySearchReturnsWhenFalse() {
-        when(publicationService.findAllBySearch(SEARCH_TERM, TEST_STRING, USER_ID)).thenReturn(List.of(artefactWithId));
-        assertEquals(HttpStatus.OK, publicationController.getAllRelevantArtefactsBySearchValue(SEARCH_TERM, TEST_STRING,
-                                                                                               USER_ID).getStatusCode(),
+        when(artefactSearchService.findAllBySearch(SEARCH_TERM, TEST_STRING, USER_ID))
+            .thenReturn(List.of(artefactWithId));
+        assertEquals(HttpStatus.OK, publicationController
+                         .getAllRelevantArtefactsBySearchValue(SEARCH_TERM, TEST_STRING, USER_ID)
+                         .getStatusCode(),
                      STATUS_CODE_MATCH);
     }
 
     @Test
     void checkGetMetadataContentReturns() {
-        when(publicationService.getMetadataByArtefactId(any(), any()))
+        when(artefactService.getMetadataByArtefactId(any(), any()))
             .thenReturn(artefactWithId);
         ResponseEntity<Artefact> unmappedBlob = publicationController
             .getArtefactMetadata(UUID.randomUUID(), USER_ID, false);
@@ -200,7 +219,7 @@ class PublicationControllerTest {
 
     @Test
     void checkGetMetadataContentReturnsAdmin() {
-        when(publicationService.getMetadataByArtefactId(any()))
+        when(artefactService.getMetadataByArtefactId(any()))
             .thenReturn(artefactWithId);
         ResponseEntity<Artefact> unmappedBlob = publicationController
             .getArtefactMetadata(UUID.randomUUID(), USER_ID, true);
@@ -213,7 +232,7 @@ class PublicationControllerTest {
 
     @Test
     void checkCountArtefactByLocationReturnsData() {
-        when(publicationService.countArtefactsByLocation()).thenReturn(COUNT_MSG);
+        when(artefactService.countArtefactsByLocation()).thenReturn(COUNT_MSG);
         ResponseEntity<String> csvString = publicationController.countByLocation();
         assertEquals(HttpStatus.OK, csvString.getStatusCode(), STATUS_CODE_MATCH);
         assertEquals(COUNT_MSG, csvString.getBody(), NOT_EQUAL_MESSAGE);
@@ -222,7 +241,7 @@ class PublicationControllerTest {
 
     @Test
     void checkGetPayloadContentReturns() {
-        when(publicationService.getPayloadByArtefactId(any(), any())).thenReturn(String.valueOf(artefactWithId));
+        when(artefactService.getPayloadByArtefactId(any(), any())).thenReturn(String.valueOf(artefactWithId));
         ResponseEntity<String> unmappedBlob =
             publicationController.getArtefactPayload(UUID.randomUUID(), USER_ID, false);
         assertEquals(HttpStatus.OK, unmappedBlob.getStatusCode(),
@@ -236,8 +255,8 @@ class PublicationControllerTest {
     void checkGetFileContentReturns() {
         String string = "Hello";
         byte[] testData = string.getBytes();
-        when(publicationService.getFlatFileByArtefactID(any(), any())).thenReturn(new ByteArrayResource(testData));
-        when(publicationService.getMetadataByArtefactId(any(), any())).thenReturn(artefactWithId);
+        when(artefactService.getFlatFileByArtefactID(any(), any())).thenReturn(new ByteArrayResource(testData));
+        when(artefactService.getMetadataByArtefactId(any(), any())).thenReturn(artefactWithId);
         ResponseEntity<Resource> flatFileBlob = publicationController.getArtefactFile(
             UUID.randomUUID(),
             USER_ID,
@@ -253,8 +272,8 @@ class PublicationControllerTest {
     void checkGetFileContentAdminReturns() {
         String string = "Hello";
         byte[] testData = string.getBytes();
-        when(publicationService.getFlatFileByArtefactID(any())).thenReturn(new ByteArrayResource(testData));
-        when(publicationService.getMetadataByArtefactId(any())).thenReturn(artefactWithId);
+        when(artefactService.getFlatFileByArtefactID(any())).thenReturn(new ByteArrayResource(testData));
+        when(artefactService.getMetadataByArtefactId(any())).thenReturn(artefactWithId);
         ResponseEntity<Resource> flatFileBlob = publicationController.getArtefactFile(
             UUID.randomUUID(),
             USER_ID,
@@ -270,7 +289,7 @@ class PublicationControllerTest {
     void checkGetArtefactsByCourtIdReturnsWhenTrue() {
         List<Artefact> artefactList = List.of(artefactWithId);
 
-        when(publicationService.findAllByLocationIdAdmin(EMPTY_FIELD, USER_ID, true)).thenReturn(artefactList);
+        when(artefactSearchService.findAllByLocationIdAdmin(EMPTY_FIELD, USER_ID, true)).thenReturn(artefactList);
         ResponseEntity<List<Artefact>> unmappedArtefact = publicationController
             .getAllRelevantArtefactsByLocationId(EMPTY_FIELD, USER_ID, true);
 
@@ -282,7 +301,7 @@ class PublicationControllerTest {
     void checkGetArtefactsByCourtIdReturnsOkWhenFalse() {
         List<Artefact> artefactList = List.of(artefactWithId);
 
-        when(publicationService.findAllByLocationIdAdmin(EMPTY_FIELD, USER_ID, false)).thenReturn(artefactList);
+        when(artefactSearchService.findAllByLocationIdAdmin(EMPTY_FIELD, USER_ID, false)).thenReturn(artefactList);
         ResponseEntity<List<Artefact>> unmappedArtefact = publicationController
             .getAllRelevantArtefactsByLocationId(EMPTY_FIELD, USER_ID, false);
 
@@ -313,7 +332,7 @@ class PublicationControllerTest {
 
     @Test
     void testDeleteArtefactReturnsOk() {
-        doNothing().when(publicationService).deleteArtefactById(any(), any());
+        doNothing().when(artefactDeleteService).deleteArtefactById(any(), any());
         assertEquals(HttpStatus.OK, publicationController.deleteArtefact(TEST_STRING, TEST_STRING).getStatusCode(),
                      STATUS_CODE_MATCH);
         assertEquals(DELETED_MESSAGE + TEST_STRING,
@@ -323,7 +342,7 @@ class PublicationControllerTest {
 
     @Test
     void testGetLocationTypeReturnsOk() {
-        when(publicationService.getLocationType(ListType.CIVIL_DAILY_CAUSE_LIST)).thenReturn(LocationType.VENUE);
+        when(artefactService.getLocationType(ListType.CIVIL_DAILY_CAUSE_LIST)).thenReturn(LocationType.VENUE);
         assertEquals(HttpStatus.OK,
                      publicationController.getLocationType(ListType.CIVIL_DAILY_CAUSE_LIST).getStatusCode(),
                      STATUS_CODE_MATCH);
@@ -358,7 +377,7 @@ class PublicationControllerTest {
 
     @Test
     void testSendNewArtefactsForSubscriptionSuccess() {
-        doNothing().when(publicationService).checkNewlyActiveArtefacts();
+        doNothing().when(artefactTriggerService).checkNewlyActiveArtefacts();
         assertThat(publicationController.sendNewArtefactsForSubscription().getStatusCode())
             .as(STATUS_CODE_MATCH)
             .isEqualTo(HttpStatus.NO_CONTENT);
@@ -366,7 +385,7 @@ class PublicationControllerTest {
 
     @Test
     void testReportNoMatchArtefactsSuccess() {
-        doNothing().when(publicationService).reportNoMatchArtefacts();
+        doNothing().when(artefactTriggerService).reportNoMatchArtefacts();
         assertThat(publicationController.reportNoMatchArtefacts().getStatusCode())
             .as(STATUS_CODE_MATCH)
             .isEqualTo(HttpStatus.NO_CONTENT);
@@ -374,7 +393,7 @@ class PublicationControllerTest {
 
     @Test
     void testDeleteExpiredArtefactsSuccess() {
-        doNothing().when(publicationService).archiveExpiredArtefacts();
+        doNothing().when(artefactDeleteService).archiveExpiredArtefacts();
         assertThat(publicationController.archiveExpiredArtefacts().getStatusCode())
             .as(STATUS_CODE_MATCH)
             .isEqualTo(HttpStatus.NO_CONTENT);
@@ -385,7 +404,7 @@ class PublicationControllerTest {
         String issuerId = UUID.randomUUID().toString();
         String artefactId = UUID.randomUUID().toString();
 
-        doNothing().when(publicationService).archiveArtefactById(artefactId, issuerId);
+        doNothing().when(artefactDeleteService).archiveArtefactById(artefactId, issuerId);
 
         ResponseEntity<String> response = publicationController.archiveArtefact(issuerId, artefactId);
